refactor: Move Stop/Fail Experiment into experiment package
This moves the StopExperiment and FailExperiment functions from the server package to
the experiment package for a better fit and to later reduce a cyclical dependency.
