+++ internal/server/server_test.go
@@ -528,64 +528,3 @@ func TestFromClusterTrial(t *testing.T) {
 		})
 	}
 }
-
-func TestStopExperiment(t *testing.T) {
-	cases := []struct {
-		desc        string
-		exp         *optimizev1beta2.Experiment
-		err         error
-		expectedOut bool
-		expectedExp *optimizev1beta2.Experiment
-	}{
-		{
-			desc: "no error",
-			exp: &optimizev1beta2.Experiment{
-				ObjectMeta: metav1.ObjectMeta{},
-			},
-			err:         nil,
-			expectedOut: false,
-			expectedExp: &optimizev1beta2.Experiment{
-				ObjectMeta: metav1.ObjectMeta{},
-			},
-		},
-		{
-			desc: "error wrong type",
-			exp: &optimizev1beta2.Experiment{
-				ObjectMeta: metav1.ObjectMeta{},
-			},
-			err: &api.Error{
-				Type: experimentsv1alpha1.ErrExperimentNameInvalid,
-			},
-			expectedOut: false,
-			expectedExp: &optimizev1beta2.Experiment{
-				ObjectMeta: metav1.ObjectMeta{},
-			},
-		},
-		{
-			desc: "error",
-			exp: &optimizev1beta2.Experiment{
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{
-						optimizev1beta2.AnnotationNextTrialURL: "111",
-					},
-				},
-			},
-			err: &api.Error{
-				Type: experimentsv1alpha1.ErrExperimentStopped,
-			},
-			expectedOut: true,
-			expectedExp: &optimizev1beta2.Experiment{
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{},
-				},
-			},
-		},
-	}
-	for _, c := range cases {
-		t.Run(c.desc, func(t *testing.T) {
-			out := StopExperiment(c.exp, c.err)
-			assert.Equal(t, c.expectedOut, out)
-			assert.Equal(t, c.expectedExp.GetAnnotations(), c.exp.GetAnnotations())
-		})
-	}
-}
